requirements.txt
flask
pillow
numpy
flask-cors
